// src/shared/service/enums/access.ts
export enum AccessEnum {
    SUPER_USER = "SUPER_USER",
    DICTIONARY = "DICTIONARY",
    DICTIONARY_USERS = "DICTIONARY_USERS",
    DICTIONARY_BRAND = "DICTIONARY_BRAND",
    DICTIONARY_MODELS = "DICTIONARY_MODELS",
    DICTIONARY_SERVICE = "DICTIONARY_SERVICE",
    DICTIONARY_CARPART = "DICTIONARY_CARPART",
    PROPOSAL = "PROPOSAL",
    PROPOSAL_ALL = "PROPOSAL_ALL",
    PROPOSAL_ADD = "PROPOSAL_ADD",
    PROPOSAL_CLIENTS = "PROPOSAL_CLIENTS",
    PROPOSAL_CARS = "PROPOSAL_CARS",
    REPORT = "REPORT",
    REPORT_EVERYDAY = "REPORT_EVERYDAY",
    REPORT_TURNOVER = "REPORT_TURNOVER",
}